added parent processing and intermediate ingredients to JSON output for ingredient-sources
